Further down the rabbit hole getting things to work.
It looks like we need the win32 version of sqlite3.dll for cordova.
Placed those and updated winmd in the DLL folder.

Moved "Util" folder from SmartSync to SmartStore, updated code to
reflect that. It makes sense that the constants and util functions exist
over in SmartStore, since they can be used/apply to both projects.

"Register" function of the smartstore proxy is "working" now, have to do
some messy conversion in winrt to take the JS stuff and make it work in
C# land.

Lots more things to get working with the javascript/salesforce plugins.